Extract code into hooks
